# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store Api List",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/v1/send_sms": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for send verification sms to user",
                "operationId": "d1247469abd2f7e94eb57030e5a6c480",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "Enter mobile number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/verify_sms_code": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to check sms code that recieved by user",
                "operationId": "14dd66daa7dd22941ebef438ae54b4e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "user mobile number",
                                        "type": "string"
                                    },
                                    "code": {
                                        "description": "recieved user sms code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/home": {
            "get": {
                "tags": [
                    "Home Page"
                ],
                "description": "get home page data",
                "operationId": "f909a31373204df98173c3befdb3ec18",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "description": "Send products id in basket to payment",
                "operationId": "8434e8cd75847412ac8573a0811cb41c",
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "product_id": 2,
                                                "count": 2
                                            },
                                            {
                                                "product_id": 3,
                                                "count": 2
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/most_sold_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "e301d6dc5b43469928cd98b298626d0d",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_viewed_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "6248a052da8441534d57f6694514bbf9",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/newest_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "659ca756353d5ebc4f0c9847187c7aca",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/cheapest_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "635d9fc1ef78cbc704c0bc743f986d01",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_expensive_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "7dc90699820c7ee8228eced0fee2bc09",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_category/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "99161e6f72634af4c026b2c092eda8cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_brand/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "a69a0aa546d1af4ee61f9a6fa3c9b9f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_details/{id}": {
            "get": {
                "tags": [
                    "Product Details"
                ],
                "description": "get product details data by product id",
                "operationId": "da09d264a3e40196f296aa1baccda669",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/save_product_comment": {
            "post": {
                "tags": [
                    "Product Details"
                ],
                "description": "save user comment for product",
                "operationId": "d185eb4577ccc45bd6f6f2b7c690b1b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "product id",
                                        "type": "integer"
                                    },
                                    "parent_id": {
                                        "description": "parent id",
                                        "type": "integer"
                                    },
                                    "body": {
                                        "description": "user comment text",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/search_product": {
            "post": {
                "tags": [
                    "Products Page"
                ],
                "description": "search product",
                "operationId": "be382728c02fd77f9523cd6684bd5c7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "User Api"
                ],
                "description": "use to signin user with recieved code",
                "operationId": "5c468f9e0f6bb6a0779c4a3f7c0f7100",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "user image",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "phone": {
                                        "description": "user phone number",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "user name",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "user address",
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "description": "user postal code",
                                        "type": "string"
                                    },
                                    "lat": {
                                        "description": "user location latitude",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "lang": {
                                        "description": "user location longitude",
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "User info"
                ],
                "operationId": "5271b95af89f9aeee0d72d7790fa1875",
                "responses": {
                    "200": {
                        "description": "It's Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth Api",
            "description": "Auth Api"
        },
        {
            "name": "Home Page",
            "description": "Home Page"
        },
        {
            "name": "Payment",
            "description": "Payment"
        },
        {
            "name": "Products Page",
            "description": "Products Page"
        },
        {
            "name": "Product Details",
            "description": "Product Details"
        },
        {
            "name": "User Api",
            "description": "User Api"
        },
        {
            "name": "User info",
            "description": "User info"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}